Moving maintains to utils and fixing it
