+++ bot.js
@@ -1,11 +1,13 @@
 // -- Requiring modules
 const chalk = require('chalk');
-const { Client, Collection, MessageEmbed } = require('discord.js');
+const discord = require('discord.js');
+const { Client, Collection, MessageEmbed } = discord;
 const { config } = require('dotenv');
 const bot = new Client({ ws: { intents: ['GUILD_PRESENCES', 'GUILD_MEMBERS', 'GUILD_MESSAGES', 'GUILDS', 'DIRECT_MESSAGES', 'GUILD_MESSAGE_REACTIONS', 'DIRECT_MESSAGE_REACTIONS'] } });
 const mongoose = require('mongoose');
 global.colours = require('./jsonFiles/colours.json');
 
+global.discord = discord;
 global.MessageEmbed = MessageEmbed;
 
 
@@ -25,7 +27,7 @@ catch (err) {
 
 ['aliases', 'commands', 'items'].forEach((x) => (bot[x] = new Collection()));
 (async () => {
-	const handlers = ['utils', 'command', 'event', 'items']
+	const handlers = ['database', 'utils', 'command', 'event', 'items']
 	for (var i = 0; i < handlers.length; i++) {
 		const handler = handlers[i];
 		try {
