+++ commands/owner/maintain.js
@@ -8,7 +8,7 @@ module.exports = {
 		},
 		aliases: []
 	},
-	async run(client, message, args, utils) {
+	async run({ bot: client, message, args, utils }) {
 		if (!process.env.owners.includes(message.author.id)) return message.channel.send('This command is limited to bot developers only.');
 
 		const command = client.commands.get(args.command);
